fix(e2e): group refactoring
