+++ e2e/support/test_support.go
@@ -37,7 +37,6 @@ import (
 	"regexp"
 	"runtime/debug"
 	"strings"
-	"sync"
 	"testing"
 	"time"
 
@@ -79,9 +78,7 @@ import (
 	"github.com/apache/camel-k/v2/pkg/cmd"
 	"github.com/apache/camel-k/v2/pkg/install"
 	"github.com/apache/camel-k/v2/pkg/platform"
-	pkgutil "github.com/apache/camel-k/v2/pkg/util"
 	v2util "github.com/apache/camel-k/v2/pkg/util"
-	"github.com/apache/camel-k/v2/pkg/util/boolean"
 	"github.com/apache/camel-k/v2/pkg/util/defaults"
 	"github.com/apache/camel-k/v2/pkg/util/kubernetes"
 	"github.com/apache/camel-k/v2/pkg/util/log"
@@ -139,9 +136,6 @@ var NoOlmOperatorImage string
 
 var testContext = context.TODO()
 var testClient client.Client
-
-var testSetupMutex = sync.Mutex{}
-var kamelInstallMutex = sync.Mutex{}
 
 func init() {
 	// This line prevents controller-runtime from complaining about log.SetLogger never being called
@@ -275,95 +269,6 @@ func Kamel(t *testing.T, ctx context.Context, args ...string) *cobra.Command {
 	return KamelWithContext(t, ctx, args...)
 }
 
-func KamelInstall(t *testing.T, ctx context.Context, namespace string, args ...string) error {
-	return KamelInstallWithID(t, ctx, platform.DefaultPlatformName, namespace, args...)
-}
-
-func KamelInstallWithID(t *testing.T, ctx context.Context, operatorID string, namespace string, args ...string) error {
-	return kamelInstallWithContext(t, ctx, operatorID, namespace, true, args...)
-}
-
-func KamelInstallWithIDAndKameletCatalog(t *testing.T, ctx context.Context, operatorID string, namespace string, args ...string) error {
-	return kamelInstallWithContext(t, ctx, operatorID, namespace, false, args...)
-}
-
-func kamelInstallWithContext(t *testing.T, ctx context.Context, operatorID string, namespace string, skipKameletCatalog bool, args ...string) error {
-	kamelInstallMutex.Lock()
-	defer kamelInstallMutex.Unlock()
-
-	var installArgs []string
-
-	installArgs = []string{"install", "-n", namespace, "--operator-id", operatorID}
-
-	if !pkgutil.StringSliceExists(args, "--build-timeout") {
-		// if --build-timeout is not explicitly passed as an argument, try to configure it
-		buildTimeout := os.Getenv("CAMEL_K_TEST_BUILD_TIMEOUT")
-		if buildTimeout == "" {
-			// default Build Timeout for tests
-			buildTimeout = "10m"
-		}
-		fmt.Printf("Setting build timeout to %s\n", buildTimeout)
-		installArgs = append(installArgs, "--build-timeout", buildTimeout)
-	}
-
-	if skipKameletCatalog {
-		installArgs = append(installArgs, "--skip-default-kamelets-setup")
-	}
-
-	logLevel := os.Getenv("CAMEL_K_TEST_LOG_LEVEL")
-	if len(logLevel) > 0 {
-		fmt.Printf("Setting log-level to %s\n", logLevel)
-		installArgs = append(installArgs, "--log-level", logLevel)
-	}
-
-	mvnCLIOptions := os.Getenv("CAMEL_K_TEST_MAVEN_CLI_OPTIONS")
-	if len(mvnCLIOptions) > 0 {
-		// Split the string by spaces
-		mvnCLIArr := strings.Split(mvnCLIOptions, " ")
-		for _, mc := range mvnCLIArr {
-			mc = strings.Trim(mc, " ")
-			if len(mc) == 0 {
-				continue
-			}
-
-			fmt.Printf("Adding maven cli option %s\n", mc)
-			installArgs = append(installArgs, "--maven-cli-option", mc)
-		}
-	}
-
-	runtimeVersion := os.Getenv("CAMEL_K_TEST_RUNTIME_VERSION")
-	if runtimeVersion != "" {
-		fmt.Printf("Setting runtime version to %s\n", runtimeVersion)
-		installArgs = append(installArgs, "--runtime-version", runtimeVersion)
-	}
-	baseImage := os.Getenv("CAMEL_K_TEST_BASE_IMAGE")
-	if baseImage != "" {
-		fmt.Printf("Setting base image to %s\n", baseImage)
-		installArgs = append(installArgs, "--base-image", baseImage)
-	}
-	opImage := os.Getenv("CAMEL_K_TEST_OPERATOR_IMAGE")
-	if opImage != "" {
-		fmt.Printf("Setting operator image to %s\n", opImage)
-		installArgs = append(installArgs, "--operator-image", opImage)
-	}
-	opImagePullPolicy := os.Getenv("CAMEL_K_TEST_OPERATOR_IMAGE_PULL_POLICY")
-	if opImagePullPolicy != "" {
-		fmt.Printf("Setting operator image pull policy to %s\n", opImagePullPolicy)
-		installArgs = append(installArgs, "--operator-image-pull-policy", opImagePullPolicy)
-	}
-	if len(os.Getenv("CAMEL_K_TEST_MAVEN_CA_PEM_PATH")) > 0 {
-		certName := "myCert"
-		secretName := "maven-ca-certs"
-		CreateSecretDecoded(t, ctx, namespace, os.Getenv("CAMEL_K_TEST_MAVEN_CA_PEM_PATH"), secretName, certName)
-		installArgs = append(installArgs, "--maven-repository", os.Getenv("KAMEL_INSTALL_MAVEN_REPOSITORIES"),
-			"--maven-ca-secret", secretName+"/"+certName)
-	}
-
-	installArgs = append(installArgs, args...)
-	installCommand := KamelWithContext(t, ctx, installArgs...)
-	return installCommand.Execute()
-}
-
 func KamelRun(t *testing.T, ctx context.Context, namespace string, args ...string) *cobra.Command {
 	return KamelRunWithID(t, ctx, platform.DefaultPlatformName, namespace, args...)
 }
@@ -488,7 +393,6 @@ func MakeWithContext(t *testing.T, rule string, args ...string) *exec.Cmd {
 	}
 
 	args = append([]string{"-C", makeDir, rule}, args...)
-
 	return exec.Command("make", args...)
 }
 
@@ -2035,67 +1939,6 @@ func PlatformByName(t *testing.T, ctx context.Context, ns string, name string) f
 	}
 }
 
-func CopyIntegrationKits(t *testing.T, ctx context.Context, ns, operatorID string) error {
-	if value, ok := os.LookupEnv("CAMEL_K_TEST_COPY_INTEGRATION_KITS"); ok && value == boolean.FalseString {
-		fmt.Println("Copy integration kits optimization is disabled")
-		return nil
-	}
-
-	opns := GetEnvOrDefault("CAMEL_K_GLOBAL_OPERATOR_NS", TestDefaultNamespace)
-
-	lst := v1.NewIntegrationKitList()
-	if err := TestClient(t).List(ctx, &lst, ctrl.InNamespace(opns)); err != nil {
-		failTest(t, err)
-	}
-	for _, kit := range lst.Items {
-		if kit.Status.Image != "" && kit.Status.Phase == v1.IntegrationKitPhaseReady {
-			copyKit := v1.IntegrationKit{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: ns,
-					Name:      kit.Name,
-					Labels:    kit.Labels,
-				},
-				Spec:   *kit.Spec.DeepCopy(),
-				Status: *kit.Status.DeepCopy(),
-			}
-
-			v1.SetAnnotation(&copyKit.ObjectMeta, v1.OperatorIDAnnotation, operatorID)
-			fmt.Printf("Copy integration kit %s from namespace %s\n", kit.Name, opns)
-			if err := CreateIntegrationKit(t, ctx, &copyKit)(); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func CopyCamelCatalog(t *testing.T, ctx context.Context, ns, operatorID string) error {
-	if value, ok := os.LookupEnv("CAMEL_K_TEST_COPY_CATALOG"); ok && value == boolean.FalseString {
-		fmt.Println("Copy catalog optimization is disabled")
-		return nil
-	}
-
-	catalogName := fmt.Sprintf("camel-catalog-%s", strings.ToLower(defaults.DefaultRuntimeVersion))
-	opns := GetEnvOrDefault("CAMEL_K_GLOBAL_OPERATOR_NS", TestDefaultNamespace)
-
-	defaultCatalog := CamelCatalog(t, ctx, opns, catalogName)()
-	if defaultCatalog != nil {
-		fmt.Printf("Copy catalog %s from namespace %s\n", catalogName, opns)
-		catalog := v1.CamelCatalog{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: ns,
-				Name:      catalogName,
-			},
-			Spec: *defaultCatalog.Spec.DeepCopy(),
-		}
-		v1.SetAnnotation(&catalog.ObjectMeta, v1.OperatorIDAnnotation, operatorID)
-		return CreateCamelCatalog(t, ctx, &catalog)()
-	}
-
-	return nil
-}
-
 func IntegrationProfileByName(t *testing.T, ctx context.Context, ns string, name string) func() *v1.IntegrationProfile {
 	return func() *v1.IntegrationProfile {
 		lst := v1.NewIntegrationProfileList()
@@ -2162,22 +2005,6 @@ func UpdateIntegrationProfile(t *testing.T, ctx context.Context, ns string, upd 
 		return nil
 	}
 	return TestClient(t).Patch(ctx, target, ctrl.RawPatch(types.MergePatchType, p))
-}
-
-func CreateCamelCatalog(t *testing.T, ctx context.Context, catalog *v1.CamelCatalog) func() error {
-	return func() error {
-		testSetupMutex.Lock()
-		defer testSetupMutex.Unlock()
-		return TestClient(t).Create(ctx, catalog)
-	}
-}
-
-func CreateIntegrationKit(t *testing.T, ctx context.Context, kit *v1.IntegrationKit) func() error {
-	return func() error {
-		testSetupMutex.Lock()
-		defer testSetupMutex.Unlock()
-		return TestClient(t).Create(ctx, kit)
-	}
 }
 
 func DeleteCamelCatalog(t *testing.T, ctx context.Context, ns, name string) func() bool {
@@ -2262,10 +2089,10 @@ func DeletePlatform(t *testing.T, ctx context.Context, ns string) func() bool {
 	}
 }
 
-func UpdatePlatform(t *testing.T, ctx context.Context, ns string, name string, upd func(ip *v1.IntegrationPlatform)) error {
-	ip := PlatformByName(t, ctx, ns, name)()
+func UpdatePlatform(t *testing.T, ctx context.Context, ns string, upd func(ip *v1.IntegrationPlatform)) error {
+	ip := PlatformByName(t, ctx, ns, platform.DefaultPlatformName)()
 	if ip == nil {
-		return fmt.Errorf("unable to locate Integration Platform %s in %s", name, ns)
+		return fmt.Errorf("unable to locate Integration Platform %s in %s", platform.DefaultPlatformName, ns)
 	}
 	target := ip.DeepCopy()
 	upd(target)
